WIP Block remove

+++ Assets/Scripts/Field/Data/BlockHealthSettings.cs
@@ -5,6 +5,6 @@ namespace Field.Data
     [CreateAssetMenu(fileName = "BlockHealthSettings", menuName = "Data/BlockHealthSettings")]
     public class BlockHealthSettings : ScriptableObject, IBlockHealthSettings
     {
-        [field: SerializeField] public float MaxValue { get; private set; }
+        [field: SerializeField] public int MaxValue { get; private set; }
     }
 }
