+++ Assets/Scripts/Stats/IHealthStats.cs
@@ -4,12 +4,12 @@ namespace Stats
 {
     public interface IHealthStats
     {
-        float MaxValue { get; }
-        float CurrentHealth { get; }
+        int MaxValue { get; }
+        int CurrentHealth { get; }
         Observable<Unit> OnHealthZero { get; }
         
         void ResetHealthStat();
-        void SetDamage(float value);
-        void AddHealth(float value);
+        void SetDamage(int value);
+        void AddHealth(int value);
     }
 }
